+++ docs/changelog.rst
@@ -8,6 +8,7 @@ Changelog
 * :bug:`8088` Fix bug where the day names are wrong on the calendar page.
 * :bug:`-` 1inchv4 swaps using the OrderFilledRfq functionality will now be properly decoded
 * :feature:`-` Defisaver automation subscription activation and deactivation events will now be properly decoded.
+* :feature:`7722` Users can now export history events in their selected currency.
 * :bug:`-` ENS transaction events involving new owner for a name/subname and address changing should now be decoded properly.
 * :feature:`7544` Gearbox protocol events and balances are now decoded and detected on all supported evm chains.
 * :feature:`-` HOP governance events will now be properly decoded.